Refactor frontend utility functions and added client testing
