+++ client/src/App.jsx
@@ -1,86 +1,7 @@
 import React, { useEffect, useMemo, useState, useCallback } from 'react';
-
-async function fetchJSON(url) {
-  const res = await fetch(url, { headers: { Accept: 'application/json' } });
-  if (!res.ok) {
-    let body = '';
-    try { body = await res.text(); } catch {}
-    throw new Error(`Failed to fetch ${url} (status ${res.status}). ${body?.slice(0, 200) || ''}`);
-  }
-  const contentType = res.headers.get('content-type') || '';
-  if (!contentType.includes('application/json')) {
-    let body = '';
-    try { body = await res.text(); } catch {}
-    throw new Error(`Expected JSON from ${url} but got '${contentType}'. Body starts: ${body.slice(0, 80)}`);
-  }
-  return res.json();
-}
-
-function ordinalize(dayNumber) {
-  const n = Number(dayNumber);
-  const mod100 = n % 100;
-  if (mod100 >= 11 && mod100 <= 13) return `${n}th`;
-  switch (n % 10) {
-    case 1: return `${n}st`;
-    case 2: return `${n}nd`;
-    case 3: return `${n}rd`;
-    default: return `${n}th`;
-  }
-}
-
-function formatDeployedAtPacific(ptString) {
-  if (!ptString || typeof ptString !== 'string') return ptString;
-  const m = ptString.match(/^(\d{4})-(\d{2})-(\d{2}) (\d{2}):(\d{2}):(\d{2}) (P[SD]T)$/);
-  if (!m) return ptString;
-  const [, yyyy, mm, dd, HH, MM, SS, tz] = m;
-  const monthNames = [
-    'January','February','March','April','May','June',
-    'July','August','September','October','November','December',
-  ];
-  const monthName = monthNames[Number(mm) - 1] || mm;
-  const day = ordinalize(Number(dd));
-  return `${monthName} ${day}, ${yyyy}, ${HH}:${MM}:${SS} ${tz}`;
-}
-
-// Convert various git remote URL formats into a clean HTTPS URL suitable for display/clicking
-function gitRemoteToHttps(remote) {
-  if (!remote || typeof remote !== 'string') return '';
-  const trimmed = remote.trim();
-
-  // Handle scp-like syntax: git@host:owner/repo.git
-  const scpLikeMatch = trimmed.match(/^[\w.-]+@([^:]+):(.+)$/);
-  if (scpLikeMatch) {
-    const host = scpLikeMatch[1];
-    let path = scpLikeMatch[2];
-    if (path.endsWith('.git')) path = path.slice(0, -4);
-    return `https://${host}/${path}`;
-  }
-
-  // Normalize common protocols to https
-  let candidate = trimmed.replace(/^git\+/, '');
-  candidate = candidate.replace(/^git:\/\//, 'https://');
-  candidate = candidate.replace(/^ssh:\/\//, 'https://');
-  candidate = candidate.replace(/^http:\/\//, 'https://');
-
-  try {
-    const u = new URL(candidate);
-    let path = u.pathname || '';
-    if (path.startsWith('/')) path = path.slice(1);
-    if (path.endsWith('.git')) path = path.slice(0, -4);
-    if (!u.hostname || !path) return '';
-    return `https://${u.hostname}/${path}`;
-  } catch {
-    // As a last resort, if it looks like host/path(.git)
-    const bare = candidate.replace(/^\/*/, '');
-    const m = bare.match(/^([^/]+)\/(.+)$/);
-    if (m) {
-      let path = m[2];
-      if (path.endsWith('.git')) path = path.slice(0, -4);
-      return `https://${m[1]}/${path}`;
-    }
-  }
-  return '';
-}
+import { fetchJSON } from './utils/fetchJSON.js';
+import { formatDeployedAtPacific } from './utils/formatDeployedAtPacific.js';
+import { gitRemoteToHttps } from './utils/gitRemoteToHttps.js';
 
 export default function App() {
   const [campaigns, setCampaigns] = useState([]);
