+++ client/src/utils/ordinalize.js
@@ -1,0 +1,13 @@
+export function ordinalize(dayNumber) {
+  const n = Number(dayNumber);
+  const mod100 = n % 100;
+  if (mod100 >= 11 && mod100 <= 13) return `${n}th`;
+  switch (n % 10) {
+    case 1: return `${n}st`;
+    case 2: return `${n}nd`;
+    case 3: return `${n}rd`;
+    default: return `${n}th`;
+  }
+}
+
+
